% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertxrf.R
\name{convertxrf}
\alias{convertxrf}
\title{Converting XRF data from kcal to micromolar}
\usage{
convertxrf(projectpath, setuppath, year, first_element, last_element)
}
\arguments{
\item{setuppath}{The name of the file containing detection limits, crystal drift, molar weights, and calibration constants.}

\item{year}{The year the drift was measured closest to when your samples were analysed.}

\item{first_element}{The name of the first column containing kcps values in the generated project dataframe.}

\item{last_element}{The name of the last column containing kcps values in the generated project  dataframe.}
}
\value{
The function creates a dataframe in the long format with the new columns "Concentration" and "Detection_limit" showing the calculated concentration and the respective detection limit.
}
\description{
This function takes your generated project dataframe created with "importxrf()" and the file containing information from the XRF machine, and converts your raw data from kcal to µM.
}
\examples{
\dontrun{
df <- convertxrf(setuppath = "/inst/extdata/xrf_setup.xlsx", year = "2019", first_element = "C", last_element = "As")
}

}
