% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{importdata}
\alias{importdata}
\title{Importing the raw data from the XRF machine (.txt file)}
\usage{
importdata(datapath)
}
\arguments{
\item{datapath}{The name of your .txt file with raw data from the XRF machine.}
}
\description{
Importing the raw data from the XRF machine (.txt file)
}
