% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{importinfo}
\alias{importinfo}
\title{Importing the Excel file with information about the samples}
\usage{
importinfo(infopath)
}
\arguments{
\item{infopath}{The name of your Excel file with necessary information about the samples.}
}
\description{
Importing the Excel file with information about the samples
}
