% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{importsetup}
\alias{importsetup}
\title{Importing the Excel file containing detection limits, crystal drift, molar weights, and calibration constants}
\usage{
importsetup(setuppath)
}
\arguments{
\item{setuppath}{The name of the file containing detection limits, crystal drift, molar weights, and calibration constants.}
}
\description{
Importing the Excel file containing detection limits, crystal drift, molar weights, and calibration constants
}
