% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importxrf.R
\name{importxrf}
\alias{importxrf}
\title{Importing and joining data file and info file}
\usage{
importxrf(datapath, infopath)
}
\arguments{
\item{datapath}{The name of your .txt file with raw data from the XRF machine.}

\item{infopath}{The name of your Excel file with necessary information about the samples.}
}
\value{
description The function creates a dataframe where your raw data and project information has been merged.
}
\description{
This function imports your files using specific utility functions from this package,  and combines them into one dataframe. You will also get a dataframe showing any samples that did    not match between your datasets (the raw data and the information file).
}
\examples{
\dontrun{
importxrf(datapath = "/inst/extdata/xrf_rawdata.txt", infopath = "/inst/extdata/xrf_projectinfo.xlsx")
# this generates a dataframe that shows up in your environment
}

}
