% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after.R
\name{show_means}
\alias{show_means}
\title{Showing the mean blanks}
\usage{
show_means(imported_data, first_element, last_element)
}
\arguments{
\item{imported_data}{The name of the dataframe created with readxrf()}

\item{first_element}{The name of the first column containing kcps values in the project dataframe.}

\item{last_element}{The name of the last column containing kcps values in the project dataframe.}
}
\value{
description The function creates a dataframe showing the mean blank values in kcps per filter type, size, and filter box number.
}
\description{

}
\examples{


}
