% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformssd.R
\name{transformssd}
\alias{transformssd}
\title{Transforming hexadecimal data from .ssd file to a readable format}
\usage{
transformssd(hex_data)
}
\arguments{
\item{hex_data}{The character string with your hexadecimal data, read into R with the readLines() function.}
}
\value{
description The function creates a dataframe showing kcps values for each element for the sample.
}
\description{
This function will transform your .txt file created from the .ssd file into a dataframe. This dataframe will contain the sample date and kcps values for each element for one sample.
}
\examples{
\dontrun{
string <- readLines("sample_hex.txt")
sample.df <- transformssd(hex_data = string)
}

}
