% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after.R
\name{widen}
\alias{widen}
\title{Widening the dataframe}
\usage{
widen(project_data)
}
\arguments{
\item{project_data}{The name of the dataframe created with convertxrf().}
}
\value{
description The function creates a dataframe that is a wider version of the one created with convertxrf().
}
\description{
This function will widen the dataframe and thus return the data to where each element is in a column.

See vignette("xrfr") for more information.
}
\examples{
\dontrun{
rawdata.df <- read_delim("xrf_rawdata.txt", delim = "\t", locale = locale(decimal_mark = ","))
projectinfo.df <- read_excel("xrf_projectinfo.xlsx")
baseinfo.df <- read_excel("xrf_setup.xlsx")

projectfile.df <- readxrf(raw_data = rawdata.df, project_info = projectinfo.df)
project.df <- convertxrf(imported_data = projectfile.df, base_info = baseinfo.df, year = "2019", first_element = "C", last_element = "As")

wideproject.df <- widen(project_data = project.df)
}

}
