% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after.R
\name{widen_above}
\alias{widen_above}
\title{Widening the data frame and removing concentrations below the detection limits}
\usage{
widen_above(project_data)
}
\arguments{
\item{project_data}{The name of the data frame created with convertxrf().}
}
\value{
description The function creates a data frame that is a wider version of the one created with convertxrf(), and where values below the detection limits are removed.
}
\description{
This function will both widen the data frame so each element has its own column, and remove concentrations that are below the detection limits.

See vignette("xrfr") for more information.
}
\examples{
\dontrun{
rawdata.df <- read_delim("xrf_rawdata.txt", delim = "\t", locale = locale(decimal_mark = ","))
projectinfo.df <- read_excel("xrf_projectinfo.xlsx")
baseinfo.df <- read_excel("xrf_setup.xlsx")

projectfile.df <- readxrf(raw_data = rawdata.df, project_info = projectinfo.df)
project.df <- convertxrf(imported_data = projectfile.df, base_info = baseinfo.df, year = "2019", first_element = "C", last_element = "As")

abovedetlim.df <- widen_above(project_data = project.df)
}

}
