% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after.R
\name{widen_means}
\alias{widen_means}
\title{Widening the dataframe and calculating the means based on two factors}
\usage{
widen_means(project_data, first_factor, second_factor = NULL)
}
\arguments{
\item{project_data}{The name of the dataframe created with convertxrf().}

\item{first_factor}{The name of the column that shows the first or only factor you want to calculate means based on, for example depth.}

\item{second_factor}{The name of the column that shows a potential second factor you want to calculate means based on, for example location.}
}
\value{
description The function creates a dataframe where the columns available are your two factors (for example location and depth) and each element.
}
\description{
This function will widen your data like widen() but also calculate the mean concentrations based on two factors, for example location and depth.

See vignette("xrfr") for more information.
}
\examples{
\dontrun{
rawdata.df <- read_delim("xrf_rawdata.txt", delim = "\t", locale = locale(decimal_mark = ","))
projectinfo.df <- read_excel("xrf_projectinfo.xlsx")
baseinfo.df <- read_excel("xrf_setup.xlsx")

projectfile.df <- readxrf(raw_data = rawdata.df, project_info = projectinfo.df)
project.df <- convertxrf(imported_data = projectfile.df, base_info = baseinfo.df, year = "2019", first_element = "C", last_element = "As")

means.df <- widen_means(project_data = project.df, first_factor = "Treatment", second_factor = "Day")
}

}
