% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after.R
\name{widen_means_above}
\alias{widen_means_above}
\title{Widening the data frame, filtering out concentrations below detection limit, and calculating the means based on two factors}
\usage{
widen_means_above(project_data, first_factor, second_factor = NULL)
}
\arguments{
\item{project_data}{The name of the data frame created with convertxrf().}

\item{first_factor}{The name of the column that shows the first or only factor you want to calculate means based on, for example depth.}

\item{second_factor}{The name of the column that shows a potential second factor you want to calculate means based on, for example location.}
}
\value{
description The function returns a data frame that shows the mean concentrations calculated from the concentrations above the detection limits for each element based on one or two factors.
}
\description{
This function will widen your data and exclude concentrations below the detection limits like widen_above(), and will also calculate the means based on your two factors like widen_means().

See vignette("xrfr") for more information.
}
\examples{
\dontrun{
rawdata.df <- read_delim("xrf_rawdata.txt", delim = "\t", locale = locale(decimal_mark = ","))
projectinfo.df <- read_excel("xrf_projectinfo.xlsx")
baseinfo.df <- read_excel("xrf_setup.xlsx")

projectfile.df <- readxrf(raw_data = rawdata.df, project_info = projectinfo.df)
project.df <- convertxrf(imported_data = projectfile.df, base_info = baseinfo.df, year = "2019", first_element = "C", last_element = "As")

meansabovedetlim.df <- widen_means_above(project_data = project.df, first_factor = "Treatment", second_factor = "Day")
}

}
