% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertxrf.R
\name{convertxrf}
\alias{convertxrf}
\title{Converting XRF data from kcal to micromolar}
\usage{
convertxrf(imported.data, base.info, year, first_element, last_element)
}
\arguments{
\item{imported.data}{The name of the dataframe created with readxrf()}

\item{base.info}{The name of the dataframe containing detection limits, crystal drift, molar weights, and calibration constants.}

\item{year}{The year the drift was measured closest to when your samples were analysed.}

\item{first_element}{The name of the first column containing kcps values in the generated project dataframe.}

\item{last_element}{The name of the last column containing kcps values in the generated project  dataframe.}
}
\value{
description The function creates a dataframe in the long format with the new columns "Concentration" and "Detection_limit" showing the calculated concentration and the respective detection limit.
}
\description{
This function takes your generated project dataframe created with readxrf() and the file containing information from the XRF machine, and converts your raw data from kcal to micromolar.

See vignette("xrfr") for more information.
}
\examples{
\dontrun{
rawdata.df <- read_delim("xrf_rawdata.txt", delim = "\t", locale = readr::locale(decimal_mark = ","))
projectinfo.df <- read_excel("xrf_projectinfo.xlsx")
baseinfo.df <- readxl::read_excel("xrf_setup.xlsx")

projectfile.df <- readxrf(raw.data = rawdata.df, project.info = projectinfo.df)

project.df <- convertxrf(imported.data = projectfile.df, base.info = baseinfo.df, year = "2019", first_element = "C", last_element = "As")
}

}
