% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxrf.R
\name{readxrf}
\alias{readxrf}
\title{Reading and joining data file and info file}
\usage{
readxrf(raw.data, project.info)
}
\arguments{
\item{raw.data}{The name of your dataframe with raw data from the XRF machine.}

\item{project.info}{The name of your dataframe with necessary information about the samples.}
}
\value{
description The function creates a dataframe where your raw data and project information has been merged.
}
\description{
This function reads your two dataframes and combines them into one. It will warn you if there are any samples that do not match between your datasets (the raw data and the information file).

See vignette("xrfr") for more information.
}
\examples{
\dontrun{
rawdata.df <- read_delim("xrf_rawdata.txt", delim = "\t", locale = readr::locale(decimal_mark = ","))
projectinfo.df <- read_excel("xrf_projectinfo.xlsx")

projectfile.df <- readxrf(raw.data = rawdata.df, project.info = projectinfo.df)
}

}
